backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg==3.1.12
psycopg-binary==3.1.12
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
